Modified Viscoelasic material defenition; added Sparsize_reduction
For viscoelastic material mu and la(lambda) will change through time so I added dt_change to reduce these changes for the sake of performance. Introduced Sparsize_reduction since the only effect of changing mu and la is on B_dyn where K is present I splitted K to two parts of sparse data which need to be multiplied by mu OR la and added to Mass matrix.  Sparsize_reduced only changes the sparse data and not the row and column storage.